docs(PPDSC-2204): restructure so build passes
