+++ site/pages/about/release-notes/components/release-badges.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Block, Flag} from 'newskit';
-import {FullRelease, ReleaseBadge} from './types';
+import {FullRelease, ReleaseBadge} from '../utils/types';
 
 const releaseBadgePresets: {[key in ReleaseBadge]: string} = {
   patch: 'flagSolidPatch',
