+++ site/pages/about/release-notes/utils/constants.ts
@@ -1,0 +1,4 @@
+export const GITHUB_URL = 'https://github.com';
+export const GITHUB_API_URL = 'https://api.github.com';
+export const REPO = 'newscorp-ghfb/newskit';
+export const JIRA_URL = `https://nidigitalsolutions.jira.com`;
